Broken into 3 modules, addressed review comments
