+++ modules/auxiliary/scanner/msmail/host_id.go
@@ -1,16 +1,5 @@
 //usr/bin/env go run "$0" "$@"; exit "$?"
 
-/*
- OWA (Outlook Webapp) is vulnerable to time-based user enumeration attacks.
- This module leverages all known, and even some lesser-known services exposed by default
- Exchange installations to enumerate users. It also targets Office 365 for error-based user enumeration.
-
- Used for gathering information about a host that may be pointed towards an Exchange or o365 tied domain
- Queries for specific DNS records related to Office 365 integration
- Attempts to extract internal domain name for onprem instance of Exchange
- Identifies services vulnerable to time-based user enumeration for onprem Exchange
- Lists password-sprayable services exposed for onprem Exchange host
-*/
 package main
 
 import (
@@ -22,8 +11,8 @@ import (
 
 func main() {
 	metadata := &module.Metadata{
-		Name:        "msmail_ident",
-		Description: "Office 365 and Exchange Enumeration",
+		Name:        "Vulnerable domain identification",
+		Description: "Identifying potentially vulnerable Exchange endpoints",
 		Authors:     []string{"poptart", "jlarose", "Vincent Yiu", "grimhacker", "Nate Power", "Nick Powers", "clee-r7"},
 		Date:        "2018-11-06",
 		Type:        "single_scanner",
@@ -46,11 +35,11 @@ func urlEnum(hostInput string) {
 	o365Domain := hostSlice[len(hostSlice)-2] + "-" + hostSlice[len(hostSlice)-1] + ".mail.protection.outlook.com"
 	addr, err := net.LookupIP(o365Domain)
 	if err != nil {
-		module.LogInfo("Domain is not using o365 resources.")
+		module.LogInfo("[-] Domain is not using o365 resources.")
 	} else if addr == nil {
 		module.LogError("error")
 	} else {
-		module.LogInfo("Domain is using o365 resources.")
+		module.LogInfo("[+] Domain is using o365 resources.")
 	}
 	asURI := "https://" + hostInput + "/Microsoft-Server-ActiveSync"
 	adURI := "https://" + hostInput + "/autodiscover/autodiscover.xml"
